Dynamic Camera
The camera currently is static has a fixed size, some global variables and logic need to be set up to control this. The size needs to adjust based on the grid size, but UI elements should not change in size but adjust when the camera zooms out. This means a global unit variable needs to be set up updated when the camera scales the UI scales too.