Screens and Popup
Brief
This section requires a lot of thinking and explaining since one this will set a precedent for the entire game going ford. Second this will directly impact how the UI will interact with the main game script.
Problem?
The real question is do I separate the menus into its own scripts and or scenes?
The simplest solution would be to keep everything together, but it would conflict with the start and update inside my main game script. The grid when launching the game should not be there until a level is selected.
The next option is to separate the scripts but not the scene. This is a good option since we can enable the script in the scene at any time, and a pause the game which will have the same menu. The downside is that we will have two-way dependencies.
The last option is that we separate all logic into their own script and scene. This would mean we would have to make another in game menu or force the player to exit to the main menu to change the level.

I believe for now the best option is the simplest keeping everything in the same script. This would let me expand the functions of the base script and ruse many functions. A select level button could be using  the same function that is used to load a level from a previous save