Name and Logo
This section will lay out my thinking regarding the name of the game and the logo/look I have in mind.
The main consideration will be the fact that I want to sell themes or looks for microtransactions, so the logo and look need to basic to allow flexible payed themes to be applied instead. This means that the core game will mostly be free outside of cosmetic themes, potentially accessing content earlier with a theme bundle might be an option in the future.
I will then demonstrate me applying the theme and trailing a level. Some wireframes in future could be generated to set a president for other themes.